+++ src/utils/readHtml.ts
@@ -6,12 +6,13 @@ export default async (htmlPath: string, panel: vscode.WebviewPanel, shutterSound
     const html = await readFile(htmlPath, 'utf-8');
     return html
         .replace(/%CSP_SOURCE%/gu, panel.webview.cspSource)
+        .replace('%SHUTTER_SOUND%', shutterSoundUri.toString())
         .replace(
-            /(src|href)="([^"]*)"/gu,
+            /(src|href)="([^http][^"]*)"/gu,
             (_, type, src) =>
-                `${type}="${panel.webview.asWebviewUri(
-                    vscode.Uri.file(path.resolve(htmlPath, '..', src))
-                )}"`
+                src.startsWith('vscode-webview-resource') ? `${type}="${src}"` :
+                    `${type}="${panel.webview.asWebviewUri(
+                        vscode.Uri.file(path.resolve(htmlPath, '..', src))
+                    )}"`
         )
-        .replace('%SHUTTER_SOUND%', shutterSoundUri.toString());
 };